backend/requirements.txt
fastapi
uvicorn
requests
googlemaps
firebase-admin
pydantic
python-dotenv
